+++ src/main/java/com/yanny/ytech/generation/DataGeneration.java
@@ -24,5 +24,6 @@ public class DataGeneration {
         generator.addProvider(event.includeServer(), new YTechDatapackProvider(packOutput, event.getLookupProvider()));
         generator.addProvider(event.includeServer(), new YTechAdvancementProvider(packOutput, event.getLookupProvider(), event.getExistingFileHelper()));
         generator.addProvider(event.includeServer(), new YTechDataMapProvider(packOutput, event.getLookupProvider()));
+        generator.addProvider(event.includeServer(), new YTechCuriosProvider(packOutput, event.getExistingFileHelper(), event.getLookupProvider()));
     }
 }